+++ seed/csharp-sdk/csharp-grpc-proto-exhaustive/read-only-memory/src/SeedApi/DataService/Requests/UpdateRequest.cs
@@ -14,7 +14,7 @@ public record UpdateRequest
     public required string Id { get; set; }
 
     [JsonPropertyName("values")]
-    public ReadOnlyMemory<double>? Values { get; set; }
+    public ReadOnlyMemory<float>? Values { get; set; }
 
     [JsonPropertyName("set_metadata")]
     public Metadata? SetMetadata { get; set; }
